refactor: 경조사 쿼리 헬퍼 함수 추출
